requirements.txt
streamlit>=1.35.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
wordcloud>=1.9.3
matplotlib>=3.8.0
pandas>=2.0.0
python-dateutil>=2.9.0
lxml>=5.0.0
Pillow>=10.0.0
newsapi-python>=0.2.7
